Drop kit encryption: ship plaintext .zip for the shared-folder model
Reverses the per-site AES encryption from #449. The ODELIA consortium is a
closed, mutually-trusting group that hands out kits through one members-only
shared folder where every site picks its own; per-site passwords were friction
without real benefit. Kits now ship as plain <SITE>_<VERSION>.zip and the sha256
manifest stays for registry verification. A shared-folder leak now means rotating
certs, so the folder must stay members-only (documented in the handbook + READMEs).

Closes #456: the deploy orchestrators glob *.zip, which is again exactly what the
build emits, so the automated-deploy breakage disappears.
